requirements.txt
requests==2.32.3
python-dotenv==1.0.1
pandas==2.2.2
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
python-crontab==3.2.0